Pass preload script source to lib/sandboxed_renderer/init.js
Don't use remote fs to load the preload script.
Load it in the native code instead and pass the script source code as an argument to the init function.
This would allow the preload script to work even if access to fs is disabled by filtering remote.require requests in rpc-server.js for security reasons.